# docs/api/api-docs.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8443
  description: Generated server url
paths:
  /api/verify/{id}:
    put:
      tags:
      - admin-dashboard-rest-controller
      summary: "If the user is a admin, he can verify another user according to his\
        \ ID"
      operationId: verify
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: User ID not found
        "200":
          description: User Banned
          content:
            application/json: {}
  /api/unverify/{id}:
    put:
      tags:
      - admin-dashboard-rest-controller
      summary: "If the user is a admin, he can unverify another user according to\
        \ his ID"
      operationId: unverify
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: User ID not found
        "200":
          description: User Banned
          content:
            application/json: {}
...
  /api/logout:
    post:
      tags:
      - login-rest-controller
      summary: Log a user out of the application
      operationId: logOut
      responses:
        "200":
          description: Log Out successful
          content:
            application/json:
              schema:
                type: object
        "400":
          description: invalid request
  /api/login:
    post:
      tags:
      - login-rest-controller
      summary: Check the credentials of a user and let them in if they are correct
      operationId: login
      parameters:
      - name: accessToken
        in: cookie
        required: false
        schema:
          type: string
      - name: refreshToken
        in: cookie
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
        required: true
      responses:
        "400":
          description: invalid credentials
        "200":
          description: Log In successful
          content:
            application/json:
              schema:
                type: object
...
  /api/profile/{username}:
    get:
      tags:
      - rest-profile-controller
      summary: Get a User
      operationId: getUserByUsername
      parameters:
      - name: username
        in: path
        required: true
        schema:
          type: string
      responses:
        "404":
          description: User not found
        "200":
          description: User Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Basic_Basic'
...
  /api/profile/followed/{username}:
    get:
      tags:
      - rest-profile-controller
      summary: Get some followed users of a User
      operationId: getFollowed
      parameters:
      - name: from
        in: query
        required: true
        schema:
          type: integer
          format: int32
      - name: size
        in: query
        required: true
        schema:
          type: integer
          format: int32
      - name: username
        in: path
        required: true
        schema:
          type: string
      responses:
        "404":
          description: User not found
        "200":
          description: Followed Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Basic_Basic'
...
  /api/bookmarks:
    get:
      tags:
      - rest-tweet-controller
      summary: Get some of the bookmarks of the current user
      operationId: getTweetsForBookmarks
      parameters:
      - name: from
        in: query
        required: true
        schema:
          type: integer
          format: int32
      - name: size
        in: query
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: Bookmarks obtained
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TweetInformation_Basic'
        "202":
          description: Bookmarks list empty
...
  /api/notifications/delete-notification:
    delete:
      tags:
      - rest-notification-controller
      summary: Deletes a notification to a user
      operationId: deleteNotification
      parameters:
      - name: idTweet
        in: query
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              type: string
      responses:
        "200":
          description: Notification Deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Notification_Basic'
        "400":
          description: Cannot delete the notification
components:
  schemas:
    TweetInformation_Basic:
      type: object
      properties:
        numComments:
          type: integer
          format: int64
        numRetweets:
          type: integer
          format: int64
        numLikes:
          type: integer
          format: int64
        colorLike:
          type: string
        colorRetweet:
          type: string
        colorBookmark:
          type: string
        isAuthorised:
          type: boolean
        urlToProfilePic:
          type: string
        urlToMedia1:
          type: string
        urlToMedia2:
          type: string
        urlToMedia3:
          type: string
        urlToMedia4:
          type: string
        tweet:
          $ref: '#/components/schemas/Tweet_Basic'
        authorised:
          type: boolean
    Tweet_Basic:
      type: object
      properties:
        id:
          type: integer
          format: int64
        user:
          $ref: '#/components/schemas/User_Basic'
        publishDate:
          type: string
          format: date-time
        likes:
          type: array
          items:
            $ref: '#/components/schemas/User_Basic'
        retweets:
          type: array
          items:
            $ref: '#/components/schemas/User_Basic'
        comments:
          type: array
          items:
            $ref: '#/components/schemas/Tweet_Basic'
        text:
          type: string
        media1:
          type: object
          properties:
            binaryStream:
              type: object
        media2:
          type: object
          properties:
            binaryStream:
              type: object
        media3:
          type: object
          properties:
            binaryStream:
              type: object
        media4:
          type: object
          properties:
            binaryStream:
              type: object
    User_Basic:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        email:
          type: string
        nickname:
          type: string
        password:
          type: string
        profilePicture:
          type: object
          properties:
            binaryStream:
              type: object
        role:
          type: string
          enum:
          - ANONYMOUS
          - USER
          - ADMIN
        followers:
          type: array
          items:
            $ref: '#/components/schemas/User_Basic'
        followed:
          type: array
          items:
            $ref: '#/components/schemas/User_Basic'
        bookmarks:
          type: array
          items:
            $ref: '#/components/schemas/Tweet_Basic'
        joinDate:
          type: string
          format: date
        type:
          type: string
        verificationCode:
          type: string
        resetPasswordToken:
          type: string
        loggedIn:
          type: boolean
        enabled:
          type: boolean
        followedNumber:
          type: integer
          format: int32
        followersNumber:
          type: integer
          format: int32
        images:
          type: array
          writeOnly: true
          items:
            type: string
...
    Basic_Basic:
      type: object
...
    Notification_Basic:
      type: object
      properties:
        id:
          type: integer
          format: int64
        tweetTrigger:
          $ref: '#/components/schemas/Tweet_Basic'
        userToNotify:
          $ref: '#/components/schemas/User_Basic'
        userWhoNotifies:
          $ref: '#/components/schemas/User_Basic'
        date:
          type: string
          format: date-time
        type:
          type: string
    LoginRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
...